+++ crates/icrate/src/Foundation/additions/mod.rs
@@ -1,5 +1,4 @@
 pub use self::comparison_result::NSComparisonResult;
-pub use self::enumerator::{NSEnumerator2, NSFastEnumeration2, NSFastEnumerator2};
 pub use self::geometry::{
     CGFloat, CGPoint, CGRect, CGSize, NSMaxXEdge, NSMaxYEdge, NSMinXEdge, NSMinYEdge, NSPoint,
     NSRect, NSRectEdge, NSRectEdgeMaxX, NSRectEdgeMaxY, NSRectEdgeMinX, NSRectEdgeMinY, NSSize,
@@ -11,20 +10,22 @@ pub use self::thread::MainThreadBound;
 #[cfg(feature = "Foundation_NSThread")]
 pub use self::thread::{is_main_thread, is_multi_threaded, MainThreadMarker};
 
-mod array;
+#[macro_use]
+mod iter;
+pub mod array;
 mod attributed_string;
 mod bundle;
 mod comparison_result;
 mod data;
-mod dictionary;
-mod enumerator;
+pub mod dictionary;
+pub mod enumerator;
 mod error;
 mod exception;
 mod geometry;
 mod number;
 mod process_info;
 mod range;
-mod set;
+pub mod set;
 mod string;
 mod thread;
 mod util;