+++ crates/objc2/src/rc/id.rs
@@ -44,6 +44,29 @@ use crate::{ffi, ClassType, Message};
 /// [`Box`]: alloc::boxed::Box
 ///
 ///
+/// # Forwarding implementations
+///
+/// Since `Id<T>` is a smart pointer, it naturally [`Deref`]s to `T`, and
+/// similarly implements [`DerefMut`] when mutable.
+///
+/// On top of this, it also forwards the implementation of a bunch of standard
+/// library traits such as [`PartialEq`], [`AsRef`], and so on, so that it
+/// becomes easy to use e.g. `Id<NSString>` as if it was just an `NSString`.
+/// (Having just `NSString` is not possible since Objective-C objects cannot
+/// live on the stack, but instead must reside on the heap).
+///
+/// Note that because of current limitations in the Rust trait system, some
+/// traits like [`Default`], [`IntoIterator`], [`FromIterator`], [`From`] and
+/// [`Into`] are not directly implementable on `NSString`; for that use-case,
+/// we instead provide the [`DefaultId`], [`IdIntoIterator`] and
+/// [`IdFromIterator`] traits, which make some of the the aforementioned
+/// traits implementable on `Id`.
+///
+/// [`DefaultId`]: crate::rc::DefaultId
+/// [`IdIntoIterator`]: crate::rc::IdIntoIterator
+/// [`IdFromIterator`]: crate::rc::IdFromIterator
+///
+///
 /// # Memory layout
 ///
 /// This is guaranteed to have the same size and alignment as a pointer to the